added data generator and updated u-net

+++ segmentation_to_h5py.py
@@ -121,7 +121,7 @@ with h5py.File(sys.argv[3], 'w') as hf:
         print(image_path)
         image = read_image(image_path)
         mask = read_image(image_pairs[image_path])[:,:,0][:,:,np.newaxis]
-        mask = np.where(mask > 0,1,0)
+        mask = np.where(mask > 0,1,0).astype(np.uint8)
         edge_image = np.uint8(cv2.Laplacian(mask,cv2.CV_64F))
         num_labels, labels_im = cv2.connectedComponents(mask)
         bounding_boxes = []
